refactor: split connector target boundaries
