--- backend/internal/api/connector_target_handlers.go
@@ -1,1113 +1,0 @@
-package api
-
-import (
-	"context"
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"math"
-	"net/http"
-	"strconv"
-	"strings"
-
-	"github.com/aipermission/aipermission/backend/internal/connectorapi"
-	"github.com/aipermission/aipermission/backend/internal/connectors"
-	"github.com/aipermission/aipermission/backend/internal/connectortargets"
-)
-
-type createConnectorTargetRequest struct {
-	ProjectID     int64          `json:"project_id"`
-	ConnectorKind string         `json:"connector_kind"`
-	Name          string         `json:"name"`
-	Config        map[string]any `json:"config,omitempty"`
-	Profile       map[string]any `json:"profile,omitempty"`
-}
-
-type createConnectorCredentialProfileRequest struct {
-	Kind      string         `json:"kind"`
-	Label     string         `json:"label"`
-	Public    map[string]any `json:"public,omitempty"`
-	Secret    map[string]any `json:"secret,omitempty"`
-	RiskLabel string         `json:"risk_label,omitempty"`
-}
-
-type updateConnectorTargetRequest struct {
-	ProjectID int64          `json:"project_id"`
-	Name      string         `json:"name"`
-	Config    map[string]any `json:"config,omitempty"`
-}
-
-type updateConnectorCredentialProfileRequest struct {
-	Kind      string         `json:"kind"`
-	Label     string         `json:"label"`
-	Public    map[string]any `json:"public,omitempty"`
-	Secret    map[string]any `json:"secret,omitempty"`
-	RiskLabel string         `json:"risk_label,omitempty"`
-}
-
-type createConnectorTargetWithProfileRequest struct {
-	Target  createConnectorTargetRequest            `json:"target"`
-	Profile createConnectorCredentialProfileRequest `json:"profile"`
-}
-
-type updateConnectorTargetWithProfileRequest struct {
-	Target  updateConnectorTargetRequest            `json:"target"`
-	Profile updateConnectorCredentialProfileRequest `json:"profile"`
-}
-
-type connectorTargetTestResponse struct {
-	TargetID      int64          `json:"target_id"`
-	ProfileID     int64          `json:"profile_id"`
-	ConnectorKind string         `json:"connector_kind"`
-	OK            bool           `json:"ok"`
-	Status        string         `json:"status"`
-	Message       string         `json:"message,omitempty"`
-	Details       map[string]any `json:"details,omitempty"`
-	DurationMS    int64          `json:"duration_ms"`
-}
-
-type connectorTargetResponse struct {
-	ID            int64            `json:"id"`
-	ProjectID     int64            `json:"project_id"`
-	ProjectName   string           `json:"project_name"`
-	ProjectSlug   string           `json:"project_slug"`
-	Ref           string           `json:"ref,omitempty"`
-	ConnectorKind string           `json:"connector_kind"`
-	Name          string           `json:"name"`
-	Config        map[string]any   `json:"config,omitempty"`
-	Status        string           `json:"status"`
-	CreatedAt     string           `json:"created_at"`
-	UpdatedAt     string           `json:"updated_at"`
-	Profiles      []profileSummary `json:"profiles,omitempty"`
-}
-
-type profileSummary struct {
-	ID                int64                         `json:"id"`
-	TargetID          int64                         `json:"target_id"`
-	Ref               string                        `json:"ref"`
-	ConnectorKind     string                        `json:"connector_kind"`
-	Kind              string                        `json:"kind"`
-	Label             string                        `json:"label"`
-	Public            map[string]any                `json:"public,omitempty"`
-	RiskLabel         string                        `json:"risk_label,omitempty"`
-	Actions           []connectors.ActionDefinition `json:"actions,omitempty"`
-	RuntimeID         int64                         `json:"runtime_id,omitempty"`
-	TransferRuntimeID int64                         `json:"transfer_runtime_id,omitempty"`
-	VaultSession      bool                          `json:"vault_session_supported"`
-	CreatedAt         string                        `json:"created_at"`
-	UpdatedAt         string                        `json:"updated_at"`
-}
-
-func validateConnectorTargetSchema(connector connectors.Connector) error {
-	return connectors.ValidateNonSecretSchema(connector.TargetSchema(), connector.Kind()+" target")
-}
-
-func validateConnectorTargetConfig(connector connectors.Connector, config map[string]any) error {
-	validator, ok := connector.(connectors.TargetConfigValidator)
-	if !ok {
-		return nil
-	}
-	return validator.ValidateTargetConfig(config)
-}
-
-func (s *Server) validateConnectorTransportConfig(ctx context.Context, store *connectortargets.Store, projectID int64, config map[string]any) error {
-	mode, _ := config["connection_mode"].(string)
-	if strings.TrimSpace(mode) != "over_ssh" {
-		return nil
-	}
-	transportTargetRef, _ := config["transport_target_ref"].(string)
-	transportTargetRef = strings.TrimSpace(transportTargetRef)
-	if transportTargetRef == "" {
-		return connectortargets.ValidationError("transport target ref is required for over_ssh")
-	}
-	if err := store.ValidateTransportProject(ctx, projectID, transportTargetRef); err != nil {
-		return err
-	}
-	kind, _, _, ok := connectortargets.ParseConnectorTargetRef(transportTargetRef)
-	if !ok {
-		return connectortargets.ErrInvalidTargetRef
-	}
-	if adapter, _ := s.connectorAPIAdapterFor(kind).(connectorapi.TCPTransportAdapter); adapter == nil {
-		return connectortargets.ValidationError(fmt.Sprintf("%s connector does not expose reviewed TCP transport", kind))
-	}
-	return nil
-}
-
-type preparedConnectorCredentialProfileInput struct {
-	Kind                string
-	Label               string
-	Public              map[string]any
-	EncryptedSecretJSON string
-	EncryptedSecretPtr  *string
-	RiskLabel           string
-}
-
-type connectorCredentialProfilePayload struct {
-	Kind      string
-	Label     string
-	Public    map[string]any
-	Secret    map[string]any
-	RiskLabel string
-}
-
-func createProfileAdapterRequest(request createConnectorCredentialProfileRequest) connectorCredentialProfilePayload {
-	return connectorCredentialProfilePayload{
-		Kind:      request.Kind,
-		Label:     request.Label,
-		Public:    request.Public,
-		Secret:    request.Secret,
-		RiskLabel: request.RiskLabel,
-	}
-}
-
-func updateProfileAdapterRequest(request updateConnectorCredentialProfileRequest) connectorCredentialProfilePayload {
-	return connectorCredentialProfilePayload{
-		Kind:      request.Kind,
-		Label:     request.Label,
-		Public:    request.Public,
-		Secret:    request.Secret,
-		RiskLabel: request.RiskLabel,
-	}
-}
-
-func (s connectorTargetHandlers) prepareConnectorCredentialProfileInput(
-	w http.ResponseWriter,
-	r *http.Request,
-	runtime *databaseRuntime,
-	connector connectors.Connector,
-	request connectorCredentialProfilePayload,
-	secretRequired bool,
-	previous *connectors.CredentialProfileView,
-	previousEncryptedSecret string,
-) (preparedConnectorCredentialProfileInput, bool) {
-	kind := strings.TrimSpace(request.Kind)
-	if !credentialKindSupported(connector, kind) {
-		writeError(w, http.StatusBadRequest, "unsupported credential kind")
-		return preparedConnectorCredentialProfileInput{}, false
-	}
-	schema, ok := credentialSchemaForKind(connector, kind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported credential kind")
-		return preparedConnectorCredentialProfileInput{}, false
-	}
-	secret, err := mergeConnectorCredentialSecrets(runtime, previousEncryptedSecret, request.Secret)
-	if err != nil {
-		writeInternalError(w)
-		return preparedConnectorCredentialProfileInput{}, false
-	}
-	if err := connectors.ValidateCredentialSchemaValues(schema.Schema, request.Public, secret, secretRequired); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return preparedConnectorCredentialProfileInput{}, false
-	}
-	public, err := s.canonicalCredentialPublic(r.Context(), runtime, connector.Kind(), kind, request.Public)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return preparedConnectorCredentialProfileInput{}, false
-	}
-	if err := connectors.ValidateCredentialSchemaValues(schema.Schema, public, secret, secretRequired); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return preparedConnectorCredentialProfileInput{}, false
-	}
-	if validator, ok := connector.(connectors.CredentialProfileValidator); ok {
-		if err := validator.ValidateCredentialProfile(kind, public, secret, previous); err != nil {
-			writeError(w, http.StatusBadRequest, err.Error())
-			return preparedConnectorCredentialProfileInput{}, false
-		}
-	}
-	prepared := preparedConnectorCredentialProfileInput{
-		Kind:      kind,
-		Label:     request.Label,
-		Public:    public,
-		RiskLabel: request.RiskLabel,
-	}
-	if secretRequired {
-		if secret == nil {
-			secret = map[string]any{}
-		}
-		encrypted, err := runtime.vault.EncryptJSON(secret)
-		if err != nil {
-			writeInternalError(w)
-			return preparedConnectorCredentialProfileInput{}, false
-		}
-		prepared.EncryptedSecretJSON = encrypted
-		prepared.EncryptedSecretPtr = &prepared.EncryptedSecretJSON
-		return prepared, true
-	}
-	if request.Secret != nil {
-		encrypted, err := runtime.vault.EncryptJSON(secret)
-		if err != nil {
-			writeInternalError(w)
-			return preparedConnectorCredentialProfileInput{}, false
-		}
-		prepared.EncryptedSecretJSON = encrypted
-		prepared.EncryptedSecretPtr = &prepared.EncryptedSecretJSON
-	}
-	return prepared, true
-}
-
-func mergeConnectorCredentialSecrets(runtime *databaseRuntime, previousEncrypted string, updates map[string]any) (map[string]any, error) {
-	if updates == nil && previousEncrypted == "" {
-		return nil, nil
-	}
-	merged := map[string]any{}
-	if previousEncrypted != "" {
-		if err := runtime.vault.DecryptJSON(previousEncrypted, &merged); err != nil {
-			return nil, err
-		}
-	}
-	for key, value := range updates {
-		if value == nil {
-			delete(merged, key)
-			continue
-		}
-		merged[key] = value
-	}
-	return merged, nil
-}
-
-func (s connectorTargetHandlers) listConnectorTargets(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	kind := strings.TrimSpace(r.URL.Query().Get("kind"))
-	store := connectortargets.NewStore(runtime.database)
-	targets, err := store.ListTargets(r.Context(), connectortargets.ListTargetsFilter{ConnectorKind: kind})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	items := make([]connectorTargetResponse, 0, len(targets))
-	for _, target := range targets {
-		items = append(items, connectorTargetToResponse(target, nil))
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"items": items})
-}
-
-func (s connectorTargetHandlers) listConnectorTargetInventory(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	kind := strings.TrimSpace(r.URL.Query().Get("kind"))
-	store := connectortargets.NewStore(runtime.database)
-	targets, err := store.ListTargets(r.Context(), connectortargets.ListTargetsFilter{ConnectorKind: kind})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	items := make([]connectorTargetResponse, 0, len(targets))
-	for _, target := range targets {
-		connector, ok := registry.Get(target.ConnectorKind)
-		if !ok {
-			writeError(w, http.StatusBadRequest, "unsupported connector kind")
-			return
-		}
-		profiles, err := store.ListCredentialProfiles(r.Context(), target.ID)
-		if err != nil {
-			handleConnectorTargetError(w, err)
-			return
-		}
-		summaries := make([]profileSummary, 0, len(profiles))
-		for _, profile := range profiles {
-			targetView := connectors.TargetView{
-				ID:            target.ID,
-				Ref:           connectortargets.ConnectorTargetRef(target.ConnectorKind, target.ID, profile.ID),
-				ConnectorKind: target.ConnectorKind,
-				Name:          target.Name,
-				Config:        target.Config,
-			}
-			profileView := connectortargets.CredentialProfileView(profile)
-			actions, err := connector.GetActionList(r.Context(), targetView, profileView)
-			if err != nil {
-				writeInternalError(w)
-				return
-			}
-			if err := connectors.ValidateActionDefinitions(actions, target.ConnectorKind+" actions"); err != nil {
-				writeInternalError(w)
-				return
-			}
-			summary := profileToSummary(profile)
-			summary.Actions = actions
-			if adapter := s.connectorLiveConsoleTargetAdapterFor(target.ConnectorKind); adapter != nil {
-				surface, surfaceErr := store.GetRuntimeSurfaceByProfile(
-					r.Context(), target.ConnectorKind, target.ID, profile.ID, adapter.LiveConsoleCapabilityKind(),
-				)
-				switch {
-				case surfaceErr == nil:
-					summary.RuntimeID = surface.ID
-					summary.VaultSession = requireSessionEnvironmentCapability(r.Context(), s.Server, runtime, surface.ID) == nil
-				case !errors.Is(surfaceErr, connectortargets.ErrRuntimeSurfaceNotFound):
-					handleConnectorTargetError(w, surfaceErr)
-					return
-				}
-			}
-			if s.connectorFileTransferAdapterFor(target.ConnectorKind) != nil {
-				surface, surfaceErr := store.GetRuntimeSurfaceByProfile(
-					r.Context(), target.ConnectorKind, target.ID, profile.ID, connectortargets.RuntimeCapabilityFileTransfer,
-				)
-				switch {
-				case surfaceErr == nil:
-					summary.TransferRuntimeID = surface.ID
-				case !errors.Is(surfaceErr, connectortargets.ErrRuntimeSurfaceNotFound):
-					handleConnectorTargetError(w, surfaceErr)
-					return
-				}
-			}
-			summaries = append(summaries, summary)
-		}
-		response := connectorTargetToResponse(target, nil)
-		response.Profiles = summaries
-		items = append(items, response)
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"items": items})
-}
-
-func (s connectorTargetHandlers) createConnectorTarget(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	var request createConnectorTargetRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(strings.TrimSpace(request.ConnectorKind))
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	if err := validateConnectorTargetSchema(connector); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	config, err := connectors.NormalizeSchemaValues(connector.TargetSchema(), request.Config)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	request.Config = config
-	if err := validateConnectorTargetConfig(connector, request.Config); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	if err := s.validateConnectorTransportConfig(r.Context(), store, request.ProjectID, request.Config); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	var target connectortargets.Target
-	err = s.withAuditedMutation(
-		r.Context(), runtime, "user", nil, 0, "connector.target.created",
-		func() any {
-			return map[string]any{
-				"project_id": target.ProjectID, "target_id": target.ID,
-				"connector_kind": target.ConnectorKind, "name": target.Name,
-			}
-		},
-		func(tx *sql.Tx) error {
-			var err error
-			target, err = connectortargets.NewTxStore(tx).CreateTarget(r.Context(), connectortargets.CreateTargetInput{
-				ProjectID: request.ProjectID, ConnectorKind: strings.TrimSpace(request.ConnectorKind),
-				Name: request.Name, Config: request.Config,
-			})
-			return err
-		},
-	)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusCreated, connectorTargetToResponse(target, nil))
-}
-
-func (s connectorTargetHandlers) createConnectorTargetWithProfile(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	var request createConnectorTargetWithProfileRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(strings.TrimSpace(request.Target.ConnectorKind))
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	if err := validateConnectorTargetSchema(connector); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	targetConfig, err := connectors.NormalizeSchemaValues(connector.TargetSchema(), request.Target.Config)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	request.Target.Config = targetConfig
-	if err := validateConnectorTargetConfig(connector, request.Target.Config); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	if err := s.validateConnectorTransportConfig(r.Context(), connectortargets.NewStore(runtime.database), request.Target.ProjectID, request.Target.Config); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	preparedProfile, ok := s.prepareConnectorCredentialProfileInput(w, r, runtime, connector, createProfileAdapterRequest(request.Profile), true, nil, "")
-	if !ok {
-		return
-	}
-	var target connectortargets.Target
-	var profile connectortargets.CredentialProfile
-	err = s.withAuditedTransaction(r.Context(), runtime, func(tx *sql.Tx, appendAudit auditAppender) error {
-		store := connectortargets.NewTxStore(tx)
-		var err error
-		target, err = store.CreateTarget(r.Context(), connectortargets.CreateTargetInput{
-			ProjectID:     request.Target.ProjectID,
-			ConnectorKind: strings.TrimSpace(request.Target.ConnectorKind),
-			Name:          request.Target.Name,
-			Config:        request.Target.Config,
-		})
-		if err != nil {
-			return err
-		}
-		if adapter := s.connectorCredentialProfileLifecycleAdapterFor(target.ConnectorKind); adapter != nil {
-			if err := adapter.BeforeCreateCredentialProfile(r.Context(), runtime, store, target); err != nil {
-				return err
-			}
-		}
-		profile, err = store.CreateCredentialProfile(r.Context(), connectortargets.CreateCredentialProfileInput{
-			TargetID:            target.ID,
-			ConnectorKind:       target.ConnectorKind,
-			Kind:                preparedProfile.Kind,
-			Label:               preparedProfile.Label,
-			Public:              preparedProfile.Public,
-			EncryptedSecretJSON: preparedProfile.EncryptedSecretJSON,
-			RiskLabel:           preparedProfile.RiskLabel,
-		})
-		if err != nil {
-			return err
-		}
-		if err := s.ensureConnectorRuntimeSurfacesForProfile(r.Context(), store, target, profile); err != nil {
-			return err
-		}
-		if err := appendAudit(tx, "user", nil, 0, "connector.target.created", map[string]any{
-			"project_id": target.ProjectID, "target_id": target.ID, "connector_kind": target.ConnectorKind, "name": target.Name,
-		}); err != nil {
-			return err
-		}
-		return appendAudit(tx, "user", nil, 0, "connector.profile.created", map[string]any{
-			"target_id": target.ID, "profile_id": profile.ID, "connector_kind": target.ConnectorKind,
-			"kind": profile.Kind, "label": profile.Label,
-		})
-	})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusCreated, connectorTargetToResponse(target, []connectortargets.CredentialProfile{profile}))
-}
-
-func (s connectorTargetHandlers) testConnectorTargetDraft(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	var request createConnectorTargetRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(strings.TrimSpace(request.ConnectorKind))
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	if err := validateConnectorTargetSchema(connector); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	config, err := connectors.NormalizeSchemaValues(connector.TargetSchema(), request.Config)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	request.Config = config
-	if err := validateConnectorTargetConfig(connector, request.Config); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	if err := s.validateConnectorTransportConfig(r.Context(), connectortargets.NewStore(runtime.database), request.ProjectID, request.Config); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if adapter := s.connectorDraftTesterFor(request.ConnectorKind); adapter != nil {
-		adapter.TestDraft(s, w, r, runtime, request)
-		return
-	}
-	writeError(w, http.StatusBadRequest, "draft test is not supported for this connector")
-}
-
-func (s connectorTargetHandlers) getConnectorTarget(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	id, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	target, err := store.GetTarget(r.Context(), id)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	profiles, err := store.ListCredentialProfiles(r.Context(), target.ID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, connectorTargetToResponse(target, profiles))
-}
-
-func (s connectorTargetHandlers) updateConnectorTarget(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	id, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	var request updateConnectorTargetRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	existing, err := store.GetTarget(r.Context(), id)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(existing.ConnectorKind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	if err := validateConnectorTargetSchema(connector); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	config, err := connectors.NormalizeSchemaValues(connector.TargetSchema(), request.Config)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	request.Config = config
-	if err := validateConnectorTargetConfig(connector, request.Config); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	if request.ProjectID == 0 {
-		request.ProjectID = existing.ProjectID
-	}
-	if err := s.validateConnectorTransportConfig(r.Context(), store, request.ProjectID, request.Config); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	release, err := runtime.vaultDelivery.acquire(r.Context())
-	if err != nil {
-		writeError(w, http.StatusRequestTimeout, "connector target update was canceled")
-		return
-	}
-	defer release()
-	var target connectortargets.Target
-	var profiles []connectortargets.CredentialProfile
-	err = s.withAuditedTransaction(r.Context(), runtime, func(tx *sql.Tx, appendAudit auditAppender) error {
-		txStore := connectortargets.NewTxStore(tx)
-		var err error
-		target, err = txStore.UpdateTarget(r.Context(), connectortargets.UpdateTargetInput{
-			ID:        id,
-			ProjectID: request.ProjectID,
-			Name:      request.Name,
-			Config:    request.Config,
-		})
-		if err != nil {
-			return err
-		}
-		profiles, err = txStore.ListCredentialProfiles(r.Context(), target.ID)
-		if err != nil {
-			return err
-		}
-		for _, profile := range profiles {
-			if err := s.ensureConnectorRuntimeSurfacesForProfile(r.Context(), txStore, target, profile); err != nil {
-				return err
-			}
-		}
-		return appendAudit(tx, "user", nil, 0, "connector.target.updated", map[string]any{
-			"project_id": target.ProjectID, "target_id": target.ID, "connector_kind": target.ConnectorKind, "name": target.Name,
-		})
-	})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if err := s.afterConnectorCredentialLifecycleChange(
-		r.Context(), runtime, target.ID, 0,
-		"connector target changed; send a fresh Vault request",
-		"connector target was updated; ask the AI to send a fresh request", false,
-	); err != nil {
-		writeInternalError(w)
-		return
-	}
-	writeJSON(w, http.StatusOK, connectorTargetToResponse(target, profiles))
-}
-
-func (s connectorTargetHandlers) updateConnectorTargetWithProfile(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	id, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	profileID, ok := parsePathInt64(w, r, "profile_id", "profile_id")
-	if !ok {
-		return
-	}
-	var request updateConnectorTargetWithProfileRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	existing, err := store.GetTarget(r.Context(), id)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(existing.ConnectorKind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	if err := validateConnectorTargetSchema(connector); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	targetConfig, err := connectors.NormalizeSchemaValues(connector.TargetSchema(), request.Target.Config)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	request.Target.Config = targetConfig
-	if err := validateConnectorTargetConfig(connector, request.Target.Config); err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	if request.Target.ProjectID == 0 {
-		request.Target.ProjectID = existing.ProjectID
-	}
-	if err := s.validateConnectorTransportConfig(r.Context(), store, request.Target.ProjectID, request.Target.Config); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	existingProfile, err := store.GetCredentialProfile(r.Context(), id, profileID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	existingProfileView := connectortargets.CredentialProfileView(existingProfile)
-	preparedProfile, ok := s.prepareConnectorCredentialProfileInput(w, r, runtime, connector, updateProfileAdapterRequest(request.Profile), request.Profile.Secret != nil, &existingProfileView, existingProfile.EncryptedSecretJSON)
-	if !ok {
-		return
-	}
-	release, err := runtime.vaultDelivery.acquire(r.Context())
-	if err != nil {
-		writeError(w, http.StatusRequestTimeout, "connector target update was canceled")
-		return
-	}
-	defer release()
-	var target connectortargets.Target
-	var profile connectortargets.CredentialProfile
-	err = s.withAuditedTransaction(r.Context(), runtime, func(tx *sql.Tx, appendAudit auditAppender) error {
-		txStore := connectortargets.NewTxStore(tx)
-		var err error
-		target, err = txStore.UpdateTarget(r.Context(), connectortargets.UpdateTargetInput{
-			ID:        id,
-			ProjectID: request.Target.ProjectID,
-			Name:      request.Target.Name,
-			Config:    request.Target.Config,
-		})
-		if err != nil {
-			return err
-		}
-		profile, err = txStore.UpdateCredentialProfile(r.Context(), connectortargets.UpdateCredentialProfileInput{
-			TargetID:            target.ID,
-			ProfileID:           profileID,
-			ConnectorKind:       target.ConnectorKind,
-			Kind:                preparedProfile.Kind,
-			Label:               preparedProfile.Label,
-			Public:              preparedProfile.Public,
-			EncryptedSecretJSON: preparedProfile.EncryptedSecretPtr,
-			RiskLabel:           preparedProfile.RiskLabel,
-		})
-		if err != nil {
-			return err
-		}
-		if err := s.ensureConnectorRuntimeSurfacesForProfile(r.Context(), txStore, target, profile); err != nil {
-			return err
-		}
-		if err := appendAudit(tx, "user", nil, 0, "connector.target.updated", map[string]any{
-			"project_id": target.ProjectID, "target_id": target.ID, "connector_kind": target.ConnectorKind, "name": target.Name,
-		}); err != nil {
-			return err
-		}
-		return appendAudit(tx, "user", nil, 0, "connector.profile.updated", map[string]any{
-			"target_id": target.ID, "profile_id": profile.ID, "connector_kind": target.ConnectorKind,
-			"kind": profile.Kind, "label": profile.Label,
-		})
-	})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if err := s.afterConnectorCredentialLifecycleChange(
-		r.Context(), runtime, target.ID, 0,
-		"connector target or credential profile changed; send a fresh Vault request",
-		"connector target or credential profile was updated; ask the AI to send a fresh request", false,
-	); err != nil {
-		writeInternalError(w)
-		return
-	}
-	writeJSON(w, http.StatusOK, connectorTargetToResponse(target, []connectortargets.CredentialProfile{profile}))
-}
-
-func (s connectorTargetHandlers) deleteConnectorTarget(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	id, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	target, err := store.GetTarget(r.Context(), id)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	release, err := runtime.vaultDelivery.acquire(r.Context())
-	if err != nil {
-		writeError(w, http.StatusRequestTimeout, "connector target deletion was canceled")
-		return
-	}
-	defer release()
-	if adapter := s.connectorTargetDeleterFor(target.ConnectorKind); adapter != nil {
-		adapter.DeleteTarget(s, w, r, runtime, target)
-		return
-	}
-	if err := s.ConnectorDeleteTargetRecord(r.Context(), runtime, target, nil); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if _, err := s.ConnectorFinalizeDeletedTarget(r.Context(), runtime, target, "connector target was deleted; ask the AI to send a fresh request", nil); err != nil {
-		writeInternalError(w)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"ok": true})
-}
-
-func (s connectorTargetHandlers) finalizeDeletedConnectorTarget(w http.ResponseWriter, r *http.Request, runtime *databaseRuntime, target connectortargets.Target, staleReason string, payload map[string]any) bool {
-	_, err := s.ConnectorFinalizeDeletedTarget(r.Context(), runtime, target, staleReason, payload)
-	if err != nil {
-		writeInternalError(w)
-		return false
-	}
-	return true
-}
-
-func (s connectorTargetHandlers) staleConnectorActionRequestsForTarget(ctx context.Context, runtime *databaseRuntime, targetID int64, profileID int64, reason string, includeRunning bool) (int64, error) {
-	if runtime == nil || runtime.database == nil || targetID < 1 {
-		return 0, nil
-	}
-	input := connectortargets.StaleActionRequestsForTargetInput{
-		TargetID: targetID, ProfileID: profileID,
-		Error:         s.redactForPersistence(ctx, runtime, reason),
-		ApprovalDrift: connectorLifecycleApprovalDrift(profileID), IncludeRunning: includeRunning,
-	}
-	var result connectortargets.StaleActionRequestsForTargetResult
-	err := s.withAuditedMutation(
-		ctx, runtime, "gateway", nil, 0, "connector_action.requests.stale",
-		func() any {
-			return map[string]any{
-				"target_id": targetID, "profile_id": profileID,
-				"request_ids": result.IDs, "affected": result.Affected,
-			}
-		},
-		func(tx *sql.Tx) error {
-			var err error
-			result, err = connectortargets.NewTxStore(tx).StaleActionRequestsForTarget(ctx, input)
-			if err == nil && result.Affected == 0 {
-				return errAuditedMutationUnchanged
-			}
-			return err
-		},
-	)
-	if errors.Is(err, errAuditedMutationUnchanged) {
-		return 0, nil
-	}
-	if err != nil {
-		return 0, err
-	}
-	return result.Affected, nil
-}
-
-func (s *Server) ensureConnectorRuntimeSurfacesForProfile(ctx context.Context, store *connectortargets.Store, target connectortargets.Target, profile connectortargets.CredentialProfile) error {
-	if store == nil {
-		return nil
-	}
-	capabilities := []string{}
-	if adapter := s.connectorLiveConsoleTargetAdapterFor(target.ConnectorKind); adapter != nil {
-		capabilities = append(capabilities, adapter.LiveConsoleCapabilityKind())
-	}
-	if s.connectorFileTransferAdapterFor(target.ConnectorKind) != nil {
-		capabilities = append(capabilities, connectortargets.RuntimeCapabilityFileTransfer)
-	}
-	seen := map[string]struct{}{}
-	for _, capability := range capabilities {
-		if _, exists := seen[capability]; exists {
-			continue
-		}
-		seen[capability] = struct{}{}
-		if _, err := store.EnsureRuntimeSurface(ctx, connectortargets.EnsureRuntimeSurfaceInput{
-			ConnectorKind:  target.ConnectorKind,
-			TargetID:       target.ID,
-			ProfileID:      profile.ID,
-			CapabilityKind: capability,
-			Label:          profile.Label,
-		}); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func connectorLifecycleApprovalDrift(profileID int64) string {
-	if profileID > 0 {
-		return "profile"
-	}
-	return "target"
-}
-
-func (s connectorTargetHandlers) runConnectorTargetOperation(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	operation := strings.TrimSpace(r.PathValue("operation"))
-	store := connectortargets.NewStore(runtime.database)
-	target, err := store.GetTarget(r.Context(), targetID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	adapter := s.connectorTargetOperationRunnerFor(target.ConnectorKind)
-	if adapter == nil {
-		writeError(w, http.StatusBadRequest, "operation is not supported for this connector")
-		return
-	}
-	adapter.RunTargetOperation(s, w, r, runtime, target, operation)
-}
-
-func (s connectorTargetHandlers) canonicalCredentialPublic(ctx context.Context, runtime *databaseRuntime, connectorKind string, credentialKind string, public map[string]any) (map[string]any, error) {
-	if adapter := s.connectorCredentialCanonicalizerFor(connectorKind); adapter != nil {
-		return adapter.CanonicalCredentialPublic(ctx, s, runtime, credentialKind, public)
-	}
-	if public == nil {
-		return map[string]any{}, nil
-	}
-	copied := make(map[string]any, len(public))
-	for key, value := range public {
-		copied[key] = value
-	}
-	return copied, nil
-}
-
-func stringConfigValue(config map[string]any, key string) string {
-	value, ok := config[key]
-	if !ok || value == nil {
-		return ""
-	}
-	return strings.TrimSpace(toString(value))
-}
-
-func intConfigValue(config map[string]any, key string, fallback int) int {
-	value, ok := config[key]
-	if !ok || value == nil {
-		return fallback
-	}
-	parsed, ok := configIntValue(value)
-	if !ok || parsed == 0 {
-		return fallback
-	}
-	return parsed
-}
-
-func configIntValue(value any) (int, bool) {
-	switch typed := value.(type) {
-	case int:
-		return typed, true
-	case int64:
-		if typed < int64(math.MinInt) || typed > int64(math.MaxInt) {
-			return 0, false
-		}
-		return int(typed), true
-	case float64:
-		if typed < float64(math.MinInt) || typed > float64(math.MaxInt) {
-			return 0, false
-		}
-		return int(typed), true
-	case json.Number:
-		parsed, err := strconv.ParseInt(string(typed), 10, strconv.IntSize)
-		if err != nil {
-			return 0, false
-		}
-		return int(parsed), true
-	default:
-		parsed, err := strconv.ParseInt(strings.TrimSpace(toString(value)), 10, strconv.IntSize)
-		if err != nil {
-			return 0, false
-		}
-		return int(parsed), true
-	}
-}
-
-func int64ConfigValue(config map[string]any, key string) int64 {
-	value, ok := config[key]
-	if !ok || value == nil {
-		return 0
-	}
-	switch typed := value.(type) {
-	case int:
-		return int64(typed)
-	case int64:
-		return typed
-	case float64:
-		return int64(typed)
-	case json.Number:
-		parsed, _ := strconv.ParseInt(string(typed), 10, 64)
-		return parsed
-	default:
-		parsed, _ := strconv.ParseInt(strings.TrimSpace(toString(value)), 10, 64)
-		return parsed
-	}
-}
-
-func toString(value any) string {
-	switch typed := value.(type) {
-	case string:
-		return typed
-	case []byte:
-		return string(typed)
-	default:
-		return ""
-	}
-}
-
-func connectorTargetToResponse(target connectortargets.Target, profiles []connectortargets.CredentialProfile) connectorTargetResponse {
-	return connectorTargetResponse{
-		ID:            target.ID,
-		ProjectID:     target.ProjectID,
-		ProjectName:   target.ProjectName,
-		ProjectSlug:   target.ProjectSlug,
-		ConnectorKind: target.ConnectorKind,
-		Name:          target.Name,
-		Config:        target.Config,
-		Status:        string(target.Status),
-		CreatedAt:     target.CreatedAt,
-		UpdatedAt:     target.UpdatedAt,
-		Profiles:      profileSummaries(profiles),
-	}
-}
-
-func profileSummaries(profiles []connectortargets.CredentialProfile) []profileSummary {
-	if profiles == nil {
-		return nil
-	}
-	items := make([]profileSummary, 0, len(profiles))
-	for _, profile := range profiles {
-		items = append(items, profileToSummary(profile))
-	}
-	return items
-}
-
-func profileToSummary(profile connectortargets.CredentialProfile) profileSummary {
-	return profileSummary{
-		ID:            profile.ID,
-		TargetID:      profile.TargetID,
-		Ref:           connectortargets.ConnectorTargetRef(profile.ConnectorKind, profile.TargetID, profile.ID),
-		ConnectorKind: profile.ConnectorKind,
-		Kind:          profile.Kind,
-		Label:         profile.Label,
-		Public:        profile.Public,
-		RiskLabel:     profile.RiskLabel,
-		CreatedAt:     profile.CreatedAt,
-		UpdatedAt:     profile.UpdatedAt,
-	}
-}
-
-func credentialKindSupported(connector connectors.Connector, kind string) bool {
-	if !connectors.ValidIdentifier(kind) {
-		return false
-	}
-	for _, schema := range connector.CredentialSchemas() {
-		if schema.Kind == kind {
-			return true
-		}
-	}
-	return false
-}
-
-func credentialSchemaForKind(connector connectors.Connector, kind string) (connectors.CredentialSchema, bool) {
-	if !connectors.ValidIdentifier(kind) {
-		return connectors.CredentialSchema{}, false
-	}
-	for _, schema := range connector.CredentialSchemas() {
-		if schema.Kind == kind {
-			return schema, true
-		}
-	}
-	return connectors.CredentialSchema{}, false
-}
-
-func handleConnectorTargetError(w http.ResponseWriter, err error) {
-	var validation connectortargets.ValidationError
-	switch {
-	case errors.Is(err, connectortargets.ErrTargetNotFound), errors.Is(err, connectortargets.ErrTargetProfileNotFound):
-		writeError(w, http.StatusNotFound, "connector target not found")
-	case errors.Is(err, connectortargets.ErrInvalidTargetRef):
-		writeError(w, http.StatusBadRequest, "invalid connector target ref")
-	case errors.As(err, &validation):
-		writeError(w, http.StatusBadRequest, validation.Error())
-	default:
-		writeInternalError(w)
-	}
-}
